requirements.txt
# Core runtime dependencies (pinned to ranges; see pyproject.toml for full extras)
numpy>=1.26
pandas>=2.1
pydantic>=2.7
jinja2>=3.1
pyyaml>=6.0
scipy>=1.11

# Optional – install extras as needed:
#   pip install "hotelling[llm]"       for LLM agents (litellm, instructor, openai)
#   pip install "hotelling[spatial]"   for GIS/raster data (geopandas, rasterio, osmnx)
#   pip install "hotelling[viz]"       for plotting (matplotlib, plotly, folium, pydeck)
#   pip install "hotelling[rl]"        for RL agents (torch, gymnasium, pettingzoo)
#   pip install "hotelling[db]"        for results DB (duckdb, pyarrow)
#   pip install "hotelling[cli]"       for CLI (typer)
#   pip install "hotelling[notebooks]" for Jupyter/marimo notebooks
#   pip install "hotelling[all]"       for everything
